Move triangulateEarClipping, marker3D, wire3D to src/plugins/lib/ui/index.js exports
